+++ src/main/java/py/com/fuentepy/appfinanzasBackend/resource/presupuesto/PresupuestoRequestNew.java
@@ -31,7 +31,7 @@ public class PresupuestoRequestNew implements Serializable {
 
     @JsonProperty("monto")
     @SerializedName("monto")
-    private Long monto;
+    private Double monto;
 
     @JsonProperty("anio")
     @SerializedName("anio")
